Factored out ssh
