+++ stacktester/openstack.py
@@ -1,13 +1,5 @@
-import time
-import socket 
-import warnings
-
 import stacktester.config
 import stacktester.nova
-
-with warnings.catch_warnings():
-    warnings.simplefilter("ignore")
-    import paramiko
 
 
 class Manager(object):
@@ -20,25 +12,3 @@ class Manager(object):
                                     self.config.nova.base_url,
                                     self.config.nova.username,
                                     self.config.nova.api_key)
-
-    def get_ssh_connection(self, host, username, password):
-        """Returns an ssh connection to the specified host"""
-        ssh_timeout = self.config.nova.ssh_timeout
-        _timeout = True
-        ssh = paramiko.SSHClient()
-        ssh.set_missing_host_key_policy(
-            paramiko.AutoAddPolicy())
-        _start_time = time.time()
-
-        while (time.time() - ssh_timeout) < _start_time:
-            try:
-                ssh.connect(host, username=username, 
-                    password=password, look_for_keys=False,
-                    timeout=ssh_timeout)
-                _timeout = False
-                break
-            except socket.error:
-                continue
-        if _timeout:
-            raise socket.error("SSH connect timed out")
-        return ssh
